Proper casting

+++ quadratic-core/src/controller/operations/cell_value.rs
@@ -30,8 +30,12 @@ impl GridController {
         let mut ops = vec![];
         let mut compute_code_ops = vec![];
         let mut data_table_ops = vec![];
-        let cell_value_rect =
-            Rect::from_numbers(sheet_pos.x, sheet_pos.y, values[0].len(), values.len());
+        let cell_value_rect = Rect::from_numbers(
+            sheet_pos.x,
+            sheet_pos.y,
+            values[0].len() as i64,
+            values.len() as i64,
+        );
         let existing_data_tables = self
             .a1_context_sheet_table_bounds(sheet_pos.sheet_id)
             .into_iter()
